Extract block verifiability utils
